fix(Types): move quacker to types
